requirements.txt
python-dotenv
azure-ai-formrecognizer==3.3.0